[+] Factorize handle search params in a service
